Split osx platform export template into multiple files
